Split base S3 APIs as separate class
